experiments/lg-t2-multi-agent/pyproject.toml
[project]
name = "lg-t2-multi-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.6",
    "langchain-community>=0.3.4",
    "langchain-core>=0.3.15",
    "langchain-openai>=0.2.5",
    "langgraph>=0.2.43",
    "pydantic>=2.9.2",
    "python-dotenv>=1.0.1",
    "tavily-python>=0.5.0",
    "langchain-tools>=0.1.34",
]
